agent/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "erica-agent"
version = "0.1.0"
description = "FastAPI agent for the Erica Windows shell"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.0",
  "pydantic-settings>=2.0",
  "pyyaml>=6.0",
  "httpx>=0.27",
  "apscheduler>=3.10",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.setuptools.packages.find]
where = ["."]
include = ["erica_agent*"]

[tool.ruff]
line-length = 100
